Created VMs for all user controls
