+++ KnittingAssistant/View/userControls/ImageArea.xaml.cs
@@ -1,4 +1,6 @@
-﻿using System.Windows.Controls;
+﻿using KnittingAssistant.ViewModel;
+using System.Windows;
+using System.Windows.Controls;
 
 namespace KnittingAssistant.View.userControls
 {
@@ -7,9 +9,19 @@ namespace KnittingAssistant.View.userControls
     /// </summary>
     public partial class ImageArea : UserControl
     {
+        private ImageAreaViewModel imageAreaViewModel;
+
         public ImageArea()
         {
             InitializeComponent();
+
+            imageAreaViewModel = new ImageAreaViewModel();
+            this.Loaded += ImageArea_Loaded;
+        }
+
+        private void ImageArea_Loaded(object sender, RoutedEventArgs e)
+        {
+            this.DataContext = imageAreaViewModel;
         }
     }
 }
