+++ KnittingAssistant/View/userControls/ToolbarArea.xaml.cs
@@ -1,4 +1,6 @@
-﻿using System.Windows.Controls;
+﻿using KnittingAssistant.ViewModel;
+using System.Windows;
+using System.Windows.Controls;
 
 namespace KnittingAssistant.View.userControls
 {
@@ -7,9 +9,20 @@ namespace KnittingAssistant.View.userControls
     /// </summary>
     public partial class ToolbarArea : UserControl
     {
+        private ToolbarAreaViewModel toolbarAreaViewModel;
+
         public ToolbarArea()
         {
             InitializeComponent();
+
+            toolbarAreaViewModel = new ToolbarAreaViewModel();
+            this.Loaded += ToolbarArea_Loaded;
         }
+
+        private void ToolbarArea_Loaded(object sender, RoutedEventArgs e)
+        {
+            this.DataContext = toolbarAreaViewModel;
+        }
+
     }
 }
