+++ KnittingAssistant/ViewModel/UserControlsVM/PropertyAreaViewModel.cs
