+++ KnittingAssistant/ViewModel/UserControlsVM/ToolbarAreaViewModel.cs
@@ -1,0 +1,121 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace KnittingAssistant.ViewModel
+{
+    public class ToolbarAreaViewModel : ViewModelBase
+    {
+        #region Dependency Property
+
+        private string m_SwitchGridIconFilename;
+        public string SwitchGridIconFilename
+        {
+            get { return m_SwitchGridIconFilename; }
+            set
+            {
+                m_SwitchGridIconFilename = value;
+                OnPropertyChanged("SwitchGridIconFilename");
+            }
+        }
+
+        private string m_SwitchGridIconToolTip;
+        public string SwitchGridIconToolTip
+        {
+            get { return m_SwitchGridIconToolTip; }
+            set
+            {
+                m_SwitchGridIconToolTip = value;
+                OnPropertyChanged("SwitchGridIconToolTip");
+            }
+        }
+
+        #endregion
+
+        #region Relay Commands
+
+        private RelayCommand loadMainImageByClickCommand;
+        public RelayCommand LoadMainImageByClickCommand
+        {
+            get
+            {
+                return loadMainImageByClickCommand ??
+                    (loadMainImageByClickCommand = new RelayCommand(obj =>
+                    {
+                        string imageFilename = imageLoader.GetImageFilename();
+                        if (imageFilename.Length > 0)
+                            LoadImageOnForm(imageFilename);
+                    }));
+            }
+        }
+
+        private RelayCommand saveImageToFileCommand;
+        public RelayCommand SaveImageToFileCommand
+        {
+            get
+            {
+                return saveImageToFileCommand ??
+                    (saveImageToFileCommand = new RelayCommand(obj =>
+                    {
+                        imageSaver.SaveImage(gridLinesVis.Value ? splitImage.GridBitmapImage : splitImage.SplittedBitmapImage);
+                        currentImageState = en_ImageStates.resultImageSaved;
+                    }));
+            }
+        }
+
+        private RelayCommand changeGridLinesVisCommand;
+        public RelayCommand ChangeGridLinesVisCommand
+        {
+            get
+            {
+                return changeGridLinesVisCommand ??
+                    (changeGridLinesVisCommand = new RelayCommand(obj =>
+                    {
+                        if (gridLinesVis != null)
+                        {
+                            Image splittedImage = new Image();
+                            if ((bool)gridLinesVis)
+                            {
+                                splittedImage.Source = splitImage.SplittedBitmapImage;
+                                SwitchGridIconFilename = "../resources/grid_on_icon_1.png";
+                                SwitchGridIconToolTip = "Показать сетку";
+                            }
+                            else
+                            {
+                                splittedImage.Source = splitImage.GridBitmapImage;
+                                SwitchGridIconFilename = "../resources/grid_off_icon_1.png";
+                                SwitchGridIconToolTip = "Скрыть сетку";
+                            }
+                            Grid.SetColumn(splittedImage, 0);
+                            Grid.SetRow(splittedImage, 0);
+                            ImageArea.mainImageContainer.Children.Add(splittedImage);
+                            gridLinesVis = !gridLinesVis;
+                        }
+                    }));
+            }
+        }
+
+        private RelayCommand openColorsManagementWindowCommand;
+        public RelayCommand OpenColorsManagementWindowCommand
+        {
+            get
+            {
+                return openColorsManagementWindowCommand ??
+                    (openColorsManagementWindowCommand = new RelayCommand(obj =>
+                    {
+                        ColorsWindow colorsWindow = new ColorsWindow();
+                        colorsWindow.ShowDialog();
+                    }));
+            }
+        }
+
+        #endregion
+
+        public ToolbarAreaViewModel()
+        {
+            SwitchGridIconFilename = "../resources/grid_off_icon_1.png";
+        }
+    }
+}
